+++ lib/clixon/clixon_xml_map.h
@@ -56,7 +56,6 @@ int xml_yang_validate_all(clicon_handle h, cxobj *xt, cxobj **xret);
 int xml_yang_validate_all_top(clicon_handle h, cxobj *xt, cxobj **xret);
 int xml2cvec(cxobj *xt, yang_stmt *ys, cvec **cvv0);
 int cvec2xml_1(cvec *cvv, char *toptag, cxobj *xp, cxobj **xt0);
-
 int xml_diff(yang_stmt *yspec, cxobj *x0, cxobj *x1, 	 
 	     cxobj ***first, size_t *firstlen, 
 	     cxobj ***second, size_t *secondlen,